requirements.txt
python-docx
pypub3
